# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Yet Another Node.js Blogg Application API",
        "description": "Yet Another Node.js Blogg Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/request": {
            "post": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "body"
                    },
                    {
                        "name": "member_id",
                        "in": "body"
                    },
                    {
                        "name": "departure_latitude",
                        "in": "body"
                    },
                    {
                        "name": "departure_longitude",
                        "in": "body"
                    },
                    {
                        "name": "departure_detail",
                        "in": "body"
                    },
                    {
                        "name": "destination_latitude",
                        "in": "body"
                    },
                    {
                        "name": "destination_longitude",
                        "in": "body"
                    },
                    {
                        "name": "destination_detail",
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/price": {
            "post": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "request_id",
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "price": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/payment": {
            "post": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "body"
                    },
                    {
                        "name": "card_number",
                        "in": "body"
                    },
                    {
                        "name": "card_holder_name",
                        "in": "body"
                    },
                    {
                        "name": "card_expiry_date",
                        "in": "body"
                    },
                    {
                        "name": "card_code",
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "paid_at": {
                                    "type": "string"
                                },
                                "request_status": {
                                    "type": "string"
                                },
                                "request_id": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/pick-up": {
            "post": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/drop-off": {
            "post": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "body"
                    },
                    {
                        "name": "trip_id",
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/get-in": {
            "post": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/trip-history": {
            "get": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "member_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "request": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/cancel": {
            "post": {
                "tags": [
                    "Request"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}